Struct rumbas::data::question::Question

pub struct Question {

    pub statement: Value<ContentAreaTranslatableString>,
    pub advice: Value<ContentAreaTranslatableString>,
    pub parts: Value<Vec<Value<QuestionPart>>>,
    pub builtin_constants: Value<BuiltinConstants>,
    pub custom_constants: Value<Vec<CustomConstant>>,
    pub variables: Value<HashMap<String, Value<VariableRepresentation>>>,
    pub variables_test: Value<VariablesTest>,
    pub functions: Value<HashMap<String, Value<Function>>>,
    pub preamble: Value<Preamble>,
    pub navigation: Value<QuestionNavigation>,
    pub extensions: Value<Extensions>,
    pub diagnostic_topic_names: Value<Vec<TranslatableString>>,
    pub resources: Value<Vec<Value<ResourcePath>>>,
    pub custom_part_types: Value<Vec<CustomPartTypeDefinitionPath>>,

}

Fields
statement: Value<ContentAreaTranslatableString>

The statement is a content area which appears at the top of the question, before any input boxes. Use the statement to set up the question and provide any information the student needs to answer it.

advice: Value<ContentAreaTranslatableString>

Advice is a content area which is shown when the student presses the Reveal button to reveal the question’s answers, or at the end of the exam. The advice area is normally used to present a worked solution to the question.

parts: Value<Vec<Value<QuestionPart>>>builtin_constants: Value<BuiltinConstants>custom_constants: Value<Vec<CustomConstant>>variables: Value<HashMap<String, Value<VariableRepresentation>>>variables_test: Value<VariablesTest>functions: Value<HashMap<String, Value<Function>>>preamble: Value<Preamble>navigation: Value<QuestionNavigation>extensions: Value<Extensions>diagnostic_topic_names: Value<Vec<TranslatableString>>

The names of the topics used in diagnostic exams that this question belongs to

resources: Value<Vec<Value<ResourcePath>>>custom_part_types: Value<Vec<CustomPartTypeDefinitionPath>>

The custom part types used in this exam

Implementations

impl Question

pub fn from_name(name: &str) -> YamlResult<Question>