Trait rumbas::data::to_numbas::ToNumbas

pub trait ToNumbas<NumbasType>: Clone + RumbasCheck {
    fn to_numbas(&self, locale: &str) -> NumbasType;

    fn to_numbas_safe(&self, locale: &str) -> NumbasResult<NumbasType> { ... }

    fn to_numbas_with_name(&self, locale: &str, _name: String) -> NumbasType { ... }
}

Required methods

fn to_numbas(&self, locale: &str) -> NumbasType

Method that converts a type to another type This method assumes that it is called by a function that is initially called from to_numbas_safe

Provided methods

fn to_numbas_safe(&self, locale: &str) -> NumbasResult<NumbasType>

Method that safely converts a type to another (probably numbas) type

fn to_numbas_with_name(&self, locale: &str, _name: String) -> NumbasType