Module rumbas::exam

Contains all the exam types

Modules

diagnostic

feedback

locale

navigation

normal

numbas_settings

question_group

timing

Structs

FileReadError

InvalidExamPathError

Enums

Exam

ExamChange

ExamDesc

ExamFileType

ExamFileTypeChange

ExamFileTypeDesc

ExamFileTypeInput

ExamFileTypeInputChange

ExamFileTypeInputDesc

ExamInput

ExamInputChange

ExamInputDesc

ParseError

Functions

convert_numbas_exam

Convert a numbas exam to rumbas data Returns the name of the exam, the resulting exam (as ExamFileType) and vectors of questions and custom part type definitions