Module rumbas::support::file_reference

Structs

ContentAreaFileString

An ContentArea string that has to be read from a file.

ContentAreaFileStringDesc

ContentAreaFileStringInput

An ContentArea string that has to be read from a file.

ContentAreaFileStringInputDesc

EmbracedJMEFileString

An embraced JME string that has to be read from a file.

EmbracedJMEFileStringDesc

EmbracedJMEFileStringInput

An embraced JME string that has to be read from a file.

EmbracedJMEFileStringInputDesc

FileString

A string that has to be read from a file.

FileStringDesc

FileStringInput

A string that has to be read from a file.

FileStringInputDesc

JMEFileString

A JME string that has to be read from a file.

JMEFileStringDesc

JMEFileStringInput

A JME string that has to be read from a file.

JMEFileStringInputDesc

Enums

AnyString

AnyStringChange

AnyStringDesc

ContentAreaFileStringChange

ContentAreaFileStringInputChange

EmbracedJMEFileStringChange

EmbracedJMEFileStringInputChange

FileStringChange

FileStringInputChange

JMEFileStringChange

JMEFileStringInputChange

Constants

FILE_PREFIX

The prefix used to specify a file reference