Module rumbas::support::default

Structs

DefaultFile

Struct used to overwrite values with defaults

DefaultFileData

Enums

DefaultExamData

DefaultExamFileType

DefaultQuestionData

DefaultQuestionFileType

Traits

DefaultFileTypeMethods

Functions

combine_exam_with_default_files

Combine an exam with all data from the default files

combine_question_with_default_files

Combine a question with all data from the default files